requirements-lock.txt
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
ffmpeg-python==0.2.0
future==1.0.0
gTTS==2.4.0
idna==3.11
lxml==6.0.2
Pillow==10.0.0
pydub==0.25.1
python-pptx==0.6.21
PyYAML==6.0.1
requests==2.32.4
tqdm==4.66.1
urllib3==2.2.3
xlsxwriter==3.2.9
